limpiando codigo

+++ app/Http/Controllers/ReplacementController.php
@@ -41,22 +41,29 @@ class ReplacementController extends Controller
 
     public function loadExcel(Request $request)
     {
-        $path = $request->file('fileExcel')->store('excels');
 
-        // $this->path =(string) $this->file_name->store("excels");
-        Log::info(storage_path('app/').$path);
-        $colections = Excel::toCollection(new RepositionImportExcel, storage_path("app//".$path));
         $items =[];
-        foreach($colections[0] as $row)
+
+        if($request->file('fileExcel'))
         {
-            $item =json_decode( json_encode($row));
-            $item =(object) $item;
-            array_push($items,$item);
+            $path = $request->file('fileExcel')->store('excels');
+            // $this->path =(string) $this->file_name->store("excels");
+            // Log::info(storage_path('app/').$path);
+            $colections = Excel::toCollection(new RepositionImportExcel, storage_path("app//".$path));
+
+            foreach($colections[0] as $row)
+            {
+                $item =json_decode( json_encode($row));
+                $item =(object) $item;
+                array_push($items,$item);
+            }
+
+            unlink(storage_path('app/'.$path));
+            // Log::info(json_encode($items));
+            // Log::info('Se elimino el archivo temporal XD');
         }
 
-        unlink(storage_path('app/'.$path));
-        Log::info(json_encode($items));
-        Log::info('Se elimino el archivo temporal XD');
+
 
         return response()->json(compact('items'));
